pyproject.toml
[build-system]
requires = ["setuptools>=82", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taijiankiller"
version = "0.1.0"
description = "用多阶段 Agent 流水线续写被腰斩网络小说的 CLI 工具"
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
  "fastapi>=0.116.1,<1",
  "deepeval>=3.9.2,<4",
  "instructor>=1.14.5,<2",
  "langchain-text-splitters>=1.1.1,<2",
  "langgraph>=1.1.3,<2",
  "lightrag-hku>=1.4.11,<2",
  "litellm>=1.82.6,<2",
  "numpy>=2.4.3,<3",
  "pydantic>=2.12.5,<3",
  "pydantic-settings>=2.13.1,<3",
  "python-multipart>=0.0.20,<1",
  "python-dotenv>=1.2.2,<2",
  "rich>=14.3.3,<15",
  "typer>=0.24.1,<1",
  "uvicorn>=0.35.0,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2,<10",
  "pytest-asyncio>=1.3.0,<2",
]

[project.scripts]
taijian = "cli.main:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["cli*", "config*", "core*", "pipeline*", "webapp*"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]
